bug correction in fitness for images

+++ src/main/java/fr/jblezoray/mygeneticalgo/dna/image/FitnessHistogramWithPatch.java
@@ -33,7 +33,7 @@ implements IFitness<X> {
     
     double sumSquaredValues = 0;
     for (int n=0; n<histogram.length; n++)
-      sumSquaredValues += n * n * histogram[n];
+      sumSquaredValues += (double)n * (double)n * (double)histogram[n];
     double rms = Math.sqrt(sumSquaredValues / (this.image.getWidth() * this.image.getHeight()));
     return 100 / rms;
   }
